code cleanup

+++ app/controllers/sessions_controller.rb
@@ -4,14 +4,12 @@ class SessionsController < ApplicationController
   #
   # @option params [String] :email the user's email address
   # @option params [String] :password the user's password
-def create
+  def create
     respond_to do |format|
       user = User.find_by_email params[:email]
       if user and user.authenticate params[:password]
         token = SecureRandom.urlsafe_base64 32, false
-        if user.sessions.count > 2
-          user.sessions.first.destroy
-        end
+        user.sessions.first.destroy if user.sessions.count > 10
         if params[:remember]
           user.sessions.create token: token, expiration: Time.now.advance(months: 1)
           cookies[:session_token] = {value: token, expires_in: 1.month, secure: Rails.env.production?}
@@ -19,11 +17,10 @@ def create
           user.sessions.create token: token, expiration: Time.now.advance(hours: 1)
           session[:session_token] = token
         end
-        format.js
       else
         flash.now[:errors] = true
-        format.js
       end
+      format.js
     end
   end
